Add GITHUB_SHA environment variable
This PR makes the GITHUB_SHA default environment variable available to use in inlineScript.
This change does not address the issue #7 since it does not pass down all the environment variables (default ones and ones set by the user). This is stopgap solution that should be superseded by a more scalable solution that addresses issue #7 completely.